backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
google-generativeai
